+++ js/data/AboutMe-data.js
@@ -1,0 +1,21 @@
+"use strict";
+
+const aboutData = [
+    {
+        aboutMe:'About Me',
+        icon:'*',
+        titleShort:'It is a long established fact that a reader will be distracted \
+                    by the readable content of a page when looking at its layout.',
+        icon: '#',
+        hello: 'Hello!,I\'M Elvish Walters,',
+        name1: 'UI/UX Disigner',
+        name2: 'Phatographer',
+        name3: 'Developer',
+        titleLong: ' It is a long established fact that a reader will be distracted \
+        by the readable content of a page when looking at its layout.The point of\
+        using Lorem Ipsum is that it has a more-or-less normal distribution of letters,\
+         as opposed to using "Content here,content here", making it look like readable English.'
+    }
+]
+export default aboutData; 
+
